feat: added functionality for watched movie
