+++ src/index.js
@@ -3,5 +3,5 @@ import './sass/index.scss';
 import './js/components/header';
 import './js/components/media';
 import './js/components/modal';
-import './js/components/movieInfoModalWindow';
+import './js/components/movieModalTemplate';
 // import './js/components/modalSecondVariant';
